Only install yay if its not there

+++ install/1-yay.sh
@@ -1,6 +1,9 @@
 sudo pacman -S --needed --noconfirm base-devel
-git clone https://aur.archlinux.org/yay-bin.git
-cd yay-bin
-makepkg -si --noconfirm
-cd ~
-rm -rf yay-bin
+
+if ! command -v yay &>/dev/null; then
+  git clone https://aur.archlinux.org/yay-bin.git
+  cd yay-bin
+  makepkg -si --noconfirm
+  cd ~
+  rm -rf yay-bin
+fi
